feat: add specs / unit tests for auth controller
