--- src/test/java/com/expense/tracker/ControllerTests/AuthControllerTests.java
@@ -1,54 +1,0 @@
-package com.expense.tracker.ControllerTests;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.test.web.servlet.MockMvc;
-
-import com.expense.tracker.Service.UserService;
-import com.expense.tracker.util.JwtUtil;
-import com.fasterxml.jackson.databind.ObjectMapper;
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-@AutoConfigureMockMvc
-public class AuthControllerTests {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @MockBean
-    private UserService userService;
-
-    @MockBean
-    private JwtUtil jwtUtil;
-
-    @MockBean
-    private AuthenticationManager authenticationManager;
-
-    @Test
-    public void testRegisterUser_Success() throws Exception {
-        Map<String, String> payload = new HashMap<>();
-        payload.put("name", "example");
-        payload.put("username", "Example123");
-        payload.put("email", "example@example.com");
-        payload.put("password", "Password@123");
-
-        String jsonPayload = new ObjectMapper().writeValueAsString(payload);
-
-        mockMvc.perform(post("/api/auth/register")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(jsonPayload))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.message").value("User Registered Successfully"));
-    }
-}
